requirements.txt
transformers
huggingface_hub
streamlit
langchain_core
langchain_community
langchain_huggingface
langchain_text_splitters
accelerate
watchdog
tqdm
pydub
bitsandbytes
git+https://github.com/openai/whisper.git 
torch
torchaudio
torchvision
streamlit-mic-recorder
ffmpeg
openai-whisper
torchaudio
xgboost
requests
groq